# openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "LumenMedia.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:15080/"
    }
  ],
  "paths": {
...
    "/api/v1/libraries": {
      "get": {
        "tags": [
          "Libraries"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LibraryDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LibraryDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LibraryDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Libraries"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLibraryRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLibraryRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLibraryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/libraries/{id}": {
      "get": {
        "tags": [
          "Libraries"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryDto"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Libraries"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLibraryRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLibraryRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLibraryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Libraries"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v1/libraries/{id}/scan": {
      "post": {
        "tags": [
          "Libraries"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/libraries/{id}/refresh-metadata": {
      "post": {
        "tags": [
          "Libraries"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/RefreshLibraryMetadataRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/RefreshLibraryMetadataRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/RefreshLibraryMetadataRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryMetadataRefreshAccepted"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryMetadataRefreshAccepted"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LibraryMetadataRefreshAccepted"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/libraries/{id}/items": {
      "get": {
        "tags": [
          "Libraries"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 50
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "title"
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "asc"
            }
          },
          {
            "name": "genre",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "year",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "watched",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfMediaItemSummary"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfMediaItemSummary"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfMediaItemSummary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtworkUrls": {
        "type": "object",
        "properties": {
          "poster": {
            "type": [
              "null",
              "string"
            ]
          },
          "backdrop": {
            "type": [
              "null",
              "string"
            ]
          },
          "logo": {
            "type": [
              "null",
              "string"
            ]
          },
          "thumb": {
            "type": [
              "null",
              "string"
            ]
          },
          "banner": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "CreateLibraryRequest": {
        "required": [
          "name",
          "type"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/LibraryType"
          },
          "paths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "settings": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LibrarySettingsDto"
              }
            ]
          }
        }
      },
...
      "EpisodeSummary": {
        "required": [
          "id",
          "seriesId",
          "seasonId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "$ref": "#/components/schemas/MediaKind"
          },
          "seriesId": {
            "type": "string",
            "format": "uuid"
          },
          "seasonId": {
            "type": "string",
            "format": "uuid"
          },
          "seasonNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "episodeNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "overview": {
            "type": [
              "null",
              "string"
            ]
          },
          "airDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "runtimeMs": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "artwork": {
            "$ref": "#/components/schemas/ArtworkUrls"
          },
          "userData": {
            "$ref": "#/components/schemas/UserDataDto"
          }
        }
      },
...
      "JobDto": {
        "required": [
          "id",
          "type",
          "state"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "$ref": "#/components/schemas/JobType"
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "progress": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "message": {
            "type": [
              "null",
              "string"
            ]
          },
          "libraryId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "startedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "finishedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "error": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "JobState": {
        "enum": [
          "Queued",
          "Running",
          "Succeeded",
          "Failed",
          "Cancelled"
        ],
        "type": "string"
      },
      "JobType": {
        "enum": [
          "ScanLibrary",
          "ImportFile",
          "FetchMetadata",
          "GenerateArtwork",
          "CleanupTranscodes"
        ],
        "type": "string"
      },
...
      "LibraryDto": {
        "required": [
          "id",
          "name",
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/LibraryType"
          },
          "paths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "itemCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "settings": {
            "$ref": "#/components/schemas/LibrarySettingsDto"
          },
          "lastScanAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "LibraryMetadataRefreshAccepted": {
        "required": [
          "libraryId",
          "mode",
          "enqueuedCount"
        ],
        "type": "object",
        "properties": {
          "libraryId": {
            "type": "string",
            "format": "uuid"
          },
          "mode": {
            "$ref": "#/components/schemas/MetadataRefreshMode"
          },
          "enqueuedCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "preferredLanguage": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "LibrarySettingsDto": {
        "type": "object",
        "properties": {
          "preferredLanguage": {
            "type": "string"
          },
          "metadataProviders": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "autoScan": {
            "type": "boolean"
          }
        }
      },
      "LibraryType": {
        "enum": [
          "Movies",
          "Series"
        ],
        "type": "string"
      },
...
      "MediaItemSummary": {
        "required": [
          "id",
          "kind",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "$ref": "#/components/schemas/MediaKind"
          },
          "title": {
            "type": "string"
          },
          "originalTitle": {
            "type": [
              "null",
              "string"
            ]
          },
          "year": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "runtimeMs": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "communityRating": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "officialRating": {
            "type": [
              "null",
              "string"
            ]
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "artwork": {
            "$ref": "#/components/schemas/ArtworkUrls"
          },
          "userData": {
            "$ref": "#/components/schemas/UserDataDto"
          },
          "addedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MediaKind": {
        "enum": [
          "Movie",
          "Series",
          "Episode"
        ],
        "type": "string"
      },
...
      "MetadataRefreshMode": {
        "enum": [
          "Missing",
          "Matched",
          "All"
        ],
        "type": "string"
      },
...
      "PagedResultOfMediaItemSummary": {
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaItemSummary"
            }
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalPages": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "RefreshLibraryMetadataRequest": {
        "type": "object",
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/MetadataRefreshMode"
          },
          "preferredLanguage": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "UpdateLibraryRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "paths": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "settings": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LibrarySettingsDto"
              }
            ]
          }
        }
      },
...
      "UserDataDto": {
        "type": "object",
        "properties": {
          "watched": {
            "type": "boolean"
          },
          "playbackPositionMs": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "nextUp": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EpisodeSummary"
              }
            ]
          }
        }
      },
...
    }
  },
...
}